+++ database/migrations/2019_08_24_015111_create_species_table.php
@@ -47,6 +47,7 @@ class CreateSpeciesTable extends Migration
             $table->json('multiplication')->nullable();
             $table->json('disadvantages')->nullable();
             $table->json('interesting_cultivar')->nullable();
+            $table->longText('maintainers_note')->nullable();
             $table->timestamps();
         });
     }